Added players ability to select a magic spell, and cast it on a selected enemy. Also has a notification to pop up when you don't have enough MP, along with the greying out and spell name and reddening the spell cost. Also added UI beeps to all battle menu button clicks, and added back buttons to the target selection and spell selection menus which go back to the main menu.
Added players ability to select a magic spell, and cast it on a selected enemy. Also has a notification to pop up when you don't have enough MP, along with the greying out and spell name and reddening the spell cost. Also added UI beeps to all battle menu button clicks, and added back buttons to the target selection and spell selection menus which go back to the main menu.