UI notifications for saving, loading, using a potion, equipping an item. Created a level up UI screen that pops up when a player levels up, showing the old levels and the new levels. If multiple people level up, a next button appears to scroll through all the players that leveled up. Prevent the player from drinking a potion if hp/mp is full.
UI notifications for saving, loading, using a potion, equipping an item. Created a level up UI screen that pops up when a player levels up, showing the old levels and the new levels. If multiple people level up, a next button appears to scroll through all the players that leveled up. Prevent the player from drinking a potion if hp/mp is full.