Added a really cool quest notification system. There is now a quest icon that directly opens the quest screen when clicked (or Q). If there is a new quest (or just starting with brand new 1st quest), the quest icon will get a cool notification symbol with a nice animation. This goes away when you look at your quest screen
Added a really cool quest notification system. There is now a quest icon that directly opens the quest screen when clicked (or Q). If there is a new quest (or just starting with brand new 1st quest), the quest icon will get a cool notification symbol with a nice animation. This goes away when you look at your quest screen