Added strength and defense potions (as well as new yellow potion effects for use in battle) which permanently increase strength/defense. Added these as drops for some harder monsters, and from some of the more difficult chests. Also added a chest to the intro that has starter stuff in it.
Added strength and defense potions (as well as new yellow potion effects for use in battle) which permanently increase strength/defense. Added these as drops for some harder monsters, and from some of the more difficult chests. Also added a chest to the intro that has starter stuff in it.